Const inputs as long

+++ backend/tests/test-clear-uint8_t-rotate.cpp
@@ -27,7 +27,7 @@ int main(void) {
   // inputs is vector of vectors
   std::vector<std::vector<uint8_t>> inputs = {{1, 2, 3, 4, 5, 6, 7, 8}};
   // const_inputs is vector (same across slots)
-  std::vector<uint8_t> const_inputs = {4};
+  std::vector<long> const_inputs = {4};
   std::vector<uint8_t> exp_values = {5, 6, 7, 8, 1, 2, 3, 4};
 
   std::vector<std::vector<uint8_t>> result =
